Devlings: This commit displays the balance changes between the new devlings and the ones included in ageless 4.15.
The changes for xp, hp, cost etc. can be seen directly.

The changes in the movetypes for the elusives and overgrowns
as well as the resistance feature for the overgrown line
(not for the nailer line) can be seen in the comments.

As for the traits, loyal-like trait & low-upkeep property have been removed

Differences to the original v4.15 devlings:
 - The three new advancement units which didn't exist back at 4.15 are included as is
 - The xp for the three units which can advance to them are lowered
 - Swapped damage and number of attacks for the melee of the offender line
 - Sneak line got undead_variation=goblin
 - Devling Hero has 6 MP, not 7MP.