Devling melee nerv horribly nerved the nailers - surely to much, finding the right role for them will be the next step with the loss of ZoC, the low upkeep ability and the HP nerve they are in effect lvl0 units , who give the xp of an lvl 1 when defeated
Two ways come to mind to balance them:
Either making them real lvl0s again
Or making them real lvl1s, which was complicated because they would have to have a different role than overgrowns.
Being the weaker and cheaper version didn't work so well because they cost the same upkeep.
With the 'low upkeep' property they were sth in-between.
Now they are still, but nerved to much, and the question is where to go on from here.

For the curser line on the other hand the migration to level 1 went fine.
They are back at 14g now (was lowered to 13 in the last release) because the have actually a good movetype.

The Overgrown got nerved a few commits before, maybe slightly to much as well, but that has to be proven,
and is dependent on the nailers outcome too.